App React component
Still render it by way of a Blaze shim.
